+++ src/components/App.jsx
@@ -1,0 +1,13 @@
+import Header from "./Header";
+import Menu from "./Menu";
+import Footer from "./Footer";
+
+export default function App() {
+  return (
+    <div className="container">
+      <Header />
+      <Menu />
+      <Footer />
+    </div>
+  );
+}
