+++ src/components/Footer.jsx
@@ -1,0 +1,21 @@
+import React from "react";
+import Order from "./Order.jsx";
+
+export default function Footer() {
+  const hour = new Date().getHours();
+  const openHour = 12;
+  const closeHour = 22;
+  const isOpen = hour >= openHour && hour <= closeHour;
+
+  return (
+    <footer>
+      {isOpen ? (
+        <Order openHour={openHour} closeHour={closeHour} />
+      ) : (
+        <p>
+          We're happy to welcome you between {openHour}:00 & {closeHour}:00
+        </p>
+      )}
+    </footer>
+  );
+}
