+++ src/components/Header.jsx
@@ -1,0 +1,7 @@
+export default function Header() {
+  return (
+    <header className="header">
+      <h1>React Fast Pizza Co. </h1>
+    </header>
+  );
+}
